DealDamageを切り出し
