+++ Assets/Scripts/Battle/Enemy.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Enemy : Actor
+{
+    protected EnemyBehaviour behaviour;
+    
+    public override void DealDamage(int value)
+    {
+        base.DealDamage(value);
+        if (value <= 0) return;
+        behaviour.DamageEffect();
+        behaviour.UpdateHealthBar();
+    }
+}
